pyproject.toml
[project]
name = "anti-uav"
version = "1.0.0"
description = "UAV Detection and Tracking using YOLO and Siamese Networks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "Anti-UAV Project"}
]
keywords = ["uav", "drone", "detection", "tracking", "yolo", "computer-vision", "deep-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.14.0",
    "albumentations>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
training = [
    "wandb>=0.15.0",
    "clearml>=1.12.0",
]
export = [
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "onnxsim>=0.4.33",
]

[project.scripts]
anti-uav-train = "scripts.train:main"
anti-uav-eval = "scripts.evaluate:main"
anti-uav-infer = "scripts.infer:main"
anti-uav-export = "scripts.export:main"

[project.urls]
Homepage = "https://anti-uav.github.io"
Repository = "https://github.com/ZhaoJ9014/Anti-UAV"
Documentation = "https://anti-uav.github.io/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "detect-secrets>=1.5.0",
    "pyright>=1.1.407",
]

[tool.uv.sources]
# Use PyTorch index for CUDA builds
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "archive",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    # Style (handled by formatter or legacy patterns)
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (common in ML code with sys.path)
    "E711",   # comparison to None (legacy pattern)
    "E712",   # comparison to True/False (legacy pattern)
    "E722",   # bare except (legacy code, gradually fix)
    "E731",   # lambda assignment (common in ML code)
    "E741",   # ambiguous variable name (legacy code uses l, I, O)
    # Bugbear (legacy patterns)
    "B007",   # unused loop control variable (legacy code)
    "B008",   # function calls in argument defaults
    "B012",   # return in finally (legacy code)
    "B904",   # raise from (legacy code)
    "B905",   # zip without strict (legacy code)
    # Comprehensions
    "C901",   # too complex
    "C408",   # unnecessary dict call (legacy code style)
    "C416",   # unnecessary dict comprehension (legacy)
    # Pyupgrade (legacy code modernization - do gradually)
    "UP008",  # super() call style
    "UP031",  # percent format
    "UP032",  # f-string
    "UP038",  # isinstance union types (requires Python 3.10+ syntax)
    # Pyflakes (legacy undefined/unused)
    "F401",   # imported but unused (re-exports in __init__.py)
    "F403",   # star imports (legacy code)
    "F821",   # undefined name (legacy code with dynamic imports)
    "F841",   # local variable assigned but never used
    # Additional bugbear
    "B006",   # mutable default argument (legacy code)
    "B020",   # loop variable override (legacy code)
    # Comprehensions
    "C414",   # list() in sorted() (legacy code)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src", "tests"]
exclude = [
    "**/__pycache__",
    ".venv",
    "archive",
    "build",
    "dist",
    "anti_uav_jittor",
    "Codes",  # Legacy code - gradually migrate
    "src/tracking",  # Legacy tracking code - needs refactoring
    "src/evaluation/toolkit",  # Legacy evaluation toolkit
    "src/detection",  # Legacy detection code
]
# Lenient settings for legacy codebase
reportMissingImports = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportGeneralTypeIssues = false
reportAttributeAccessIssue = false
reportCallIssue = false
reportArgumentType = false
reportUndefinedVariable = "warning"
